requirements.txt
pyrit
python-dotenv